# perf(tokenizer): Make de-tokenize ~50% faster