--- a/Cargo.toml
+++ b/Cargo.toml
@@ -89,7 +89,7 @@ insta.opt-level = 3
 # release level optimizations otherwise everything feels slow
 opt-level = 3
 
-#[profile.release]
+[profile.release]
 # These make the build much slower but shrink the binary, and could help performance
-#codegen-units = 1
-#lto = true
+codegen-units = 1
+lto = true