--- a/lib/llm/src/preprocessor.rs
+++ b/lib/llm/src/preprocessor.rs
@@ -1,17 +1,5 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 //! The Preprocessor consists of the following modules
 //!
@@ -226,13 +214,12 @@ impl OpenAIPreprocessor {
                         TextInput::Batch(texts) => {
                             let token_batches: Vec<Vec<u32>> = texts
                                 .par_iter()
-                                .filter_map(|text| {
+                                .map(|text| {
                                     self.tokenizer
                                         .encode(text)
-                                        .ok()
                                         .map(|encoded| encoded.token_ids().to_vec())
                                 })
-                                .collect();
+                                .collect::<Result<Vec<_>>>()?;
                             builder.batch_token_ids(Some(token_batches));
                             builder.token_ids(vec![]);
                         }