- Add benchmarks for tokenize and de-tokenize
- Remove the duplicate de-tokenization, that's the big win
- Remove allocations, memory copies and memory moves in the tokenize/de-tokenize fast paths.

On a single core of my local machine I get:
- tokenizer: 2ms for 8k sequence, so about 3.8 MiB/s. This is as fast as it goes, all the time is spent in `tokenizers` dependency.
- de-tokenizer: Down from ~170ms to ~85ms for 34 tokens, about 400 bytes/s. A big speedup but still slow compared to tokenize. The difficult part is de-tokenize happens a token at a time.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

* **New Features**
  * Added benchmarking for tokenizer encoding and decoding, including integration with Criterion for performance measurement.

* **Refactor**
  * Unified encoding representation by introducing an enum for encoding types.
  * Simplified and clarified decoding logic, including improved memory allocation and early returns for certain conditions.
  * Updated all usages to access token IDs through a method rather than direct field access for consistency.

* **Bug Fixes**
  * Improved error handling and reporting in tokenizer operations.

* **Tests**
  * Updated tests to use the new token ID access method, ensuring compatibility with recent refactoring.

* **Chores**
  * Updated configuration files to support new benchmarks and streamline release profile settings.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->